+++ photon-model-adapters/azure/src/test/java/com/vmware/photon/controller/model/adapters/azure/enumeration/TestAzureEnumerationTask.java
@@ -256,7 +256,6 @@ public class TestAzureEnumerationTask extends BasicReusableHostTestCase {
     private void runEnumeration() throws Throwable {
         ResourceEnumerationTaskState enumerationTaskState = new ResourceEnumerationTaskState();
 
-        enumerationTaskState.computeDescriptionLink = this.computeHost.descriptionLink;
         enumerationTaskState.parentComputeLink = this.computeHost.documentSelfLink;
         enumerationTaskState.enumerationAction = EnumerationAction.START;
         enumerationTaskState.adapterManagementReference = UriUtils