+++ photon-model-adapters/gcp/src/test/java/com/vmware/photon/controller/model/adapters/gcp/enumeration/TestGCPEnumerationAtScale.java
@@ -132,8 +132,7 @@ public class TestGCPEnumerationAtScale extends BasicReusableHostTestCase {
                 return;
             }
 
-            ResourceEnumerationTaskState enumTask = createResourceEnumerationTask(this.outPool.documentSelfLink,
-                    this.computeHost.descriptionLink, this.computeHost.documentSelfLink,
+            ResourceEnumerationTaskState enumTask = createResourceEnumerationTask(this.outPool.documentSelfLink, this.computeHost.documentSelfLink,
                     GCPEnumerationAdapterService.SELF_LINK, this.isMock, this.computeHost.tenantLinks);
             enumerateResources(this.host, null, enumTask, TEST_CASE_BASELINE_VMS);