+++ photon-model-adapters/vsphere/adapter/src/test/java/com/vmware/photon/controller/model/adapters/vsphere/TestVSphereEnumerationTask.java
@@ -95,7 +95,6 @@ public class TestVSphereEnumerationTask extends BaseVSphereAdapterTest {
     private void doRefresh() throws Throwable {
         ResourceEnumerationTaskState task = new ResourceEnumerationTaskState();
         task.adapterManagementReference = this.computeHost.adapterManagementReference;
-        task.computeDescriptionLink = this.computeHostDescription.documentSelfLink;
 
         task.isMockRequest = isMock();
         task.enumerationAction = EnumerationAction.REFRESH;